Redesign the RL script for Nexus nodes.
(1) Remove RL-seed. The functionality is integrated in lisp-provision.py.
(2) Simplify RL. It essentially has two semantic lines, a call to
    lisp-provision.py and a call to RESTART-LISP.
(3) Add lisp-provsion.py that provisions a nexus LISP node for client use
    either within a container or natively as well as a LISP seed node
    either within a container or natively.

This allows you to remove the lisp.config file without rebooting the system
or destroying a container. When this is done, a new lisp.config is provisioned
with dynamically created EIDs (for client modes) or user supplied EIDs (for
seed nodes).